backend/requirements-api.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
joblib==1.5.3
numpy==2.4.3
scikit-learn==1.8.0
xgboost==3.2.0
shap==0.48.0
torch==2.10.0

PyJWT==2.8.0
